PR 57870, Trac #24
For tagflash wavecals, the input header gives the nominal times when the
wavecal lamp turned on and off.  When calcos looks for the actual turn-on
and turn-off times, it only looks within the nominal interval, but extended
at the start and finish by 2.0 seconds (one of the arguments to
ConcurrentWavecal).  Limiting the range of time searched for wavecal lamp
flashes is intended to reduce false positives (e.g. from FUV bursts).  The
reason for searching over a wider range than from nominal turn-on to turn-off
is that there can be a delay between the command to turn the lamp on or off
and the actual turn-on or turn-off.  Currently the nominal time interval is
extended by 2.0 seconds at both turn-on and turn-off, which is probably OK
at turn-on but is apparently too short at turn-off.  An argument should be
added to ConcurrentWavecal to specify how much the search range should be
extended past the nominal turn-off time, with a default value of 4.0 seconds.
                                                                                
The location of a wavecal spectrum in the cross-dispersion direction is
found by collapsing the data along the dispersion axis, boxcar smoothing,
and looking for the peak value over a range centered on the nominal
location of the spectrum.  The current range for FUV is 100 pixels, and for
some data where both the science and wavecal aperture were illuminated,
calcos found the science spectrum because it was brighter than the wavecal
spectrum.  The range for FUV should be reduced to 50; this is in ttFindFUV
in wavecal.py.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@709 fe389314-cf27-0410-b35b-8c050e845b92