Other than differences in syntax, the following changes were made to ccos.c: y and dq were removed from the calling sequence of getstartstop because they are not used. In extract2DBand, the loops over i and k were interchanged, and the computation of j (nearest integer to y) has been taken out of the inner loop.  The reason was to ensure that each pixel in the cross-dispersion direction would be copied out exactly once, in case there might be either skipped or duplicated pixels due to roundoff when computing nearest integer. In smoothBackground, scr was allocated to be of length length+width * sizeof (float), but it should have been (length+width) * sizeof (float).
The only change to calcosparam.py is the version string (v 1.2).

ccos.getstartstop is called in burst.py and concurrent.py, so these
were changed to fix the calling sequence of getstartstop.
In concurrent.py, "NUVC" is now excluded from segment_list
for data take with G230L, cenwave 3360.

A utility function getColCopy was added to cosutil.py.  This function
gets a column from a table, but it copies the column to a new array
to ensure that the array that is returned is contiguous, aligned, and
in the native byte order.  This is called in timetag.py for cases where
the array may be passed more than once to functions in ccos.c, to avoid
making a copy each time.  Another change to timetag.py is in the loops
over time intervals for computing deadtime and thermal distortion.
The first time interval now starts with the first time in the events
table, rather than at zero (the first time should be zero, but
currently it isn't).


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@866 fe389314-cf27-0410-b35b-8c050e845b92