### Changes to calcos.py:
Inherit from object.

The product name is now converted to lower case after reading it from
the association table.

In allScience, modified the test before calling concatenateSpectra;
the code was changed from:
        if tagflash:
            self.concatenateSpectra ("tagflash")
to:
        if tagflash and obs.switches["wavecorr"] == "PERFORM":
            self.concatenateSpectra ("tagflash")

### Change to concurrent.py:

Inherit from object.

In processConcurrentWavecal, if wavecorr is not set to PERFORM, initWavecal
is now called before returning.  This is necessary in order to copy the
pixel coordinates to the XFULL and YFULL columns.

### Change to extract.py:

In doFluxCorr, the flux correction was previously done by dividing by
the sensitivity pixel by pixel.  Now the sensitivity is interpolated
as a function of wavelength at each pixel.

### Change to fpavg.py:

Inherit from object.

Some of the code in calcos used a one-element array for the second
argument of the numarray.where function, in order to prevent the
returned array from being converted to a different data type.  This
now results in an error because the two arrays are not the same
length.  The code was changed from:
        bad = N.zeros (1, dtype=N.int16)
        bad[0] = DQ_OUT_OF_BOUNDS
        maxdq = N.where (sumweight == 0., bad, maxdq)
to:
        maxdq = N.where (sumweight == 0., DQ_OUT_OF_BOUNDS, maxdq)

### Changes to timetag.py:

In filterByPulseHeight, there was a bug in the test for the number
of events rejected because the pulse height was out of range.  The
code was changed from:
    rejected = N.nonzero (dq & DQ_PH_LOW)
    rejected = N.nonzero (dq & DQ_PH_HIGH)
to:
    rejected = N.nonzero (dq & DQ_PH_LOW)[0]
    rejected = N.nonzero (dq & DQ_PH_HIGH)[0]

In computeThermalParam, for each time interval, a warning is now printed
if either stim is not found.  Another warning will be printed if neither
stim was found in any time interval.

In thermalDistortion, there was a bug in the test regarding whether any
change actually needed to be applied.  The code was changed ('and' was
replaced with 'or') from:
        if x0[n] != 0. and xslope[n] != 1. and \
           y0[n] != 0. and yslope[n] != 1.:
            x[i:j] = x0[n] + x[i:j] * xslope[n]
            y[i:j] = y0[n] + y[i:j] * yslope[n]
            actually_done = 1
to:
        if x0[n] != 0. or xslope[n] != 1. or \
           y0[n] != 0. or yslope[n] != 1.:
            x[i:j] = x0[n] + x[i:j] * xslope[n]
            y[i:j] = y0[n] + y[i:j] * yslope[n]
            actually_done = 1

In writeImages, the output or outcounts files will now only be written
if their names are not None.  This change was made so it will be possible
to use this function in another program to recreate either or both files
(but primarily the temporary _counts file) if they don't already exist.

### Change to wavecal.py:

In xcStat, a warning will now be printed if the peak in the cross
correlation is at an endpoint of the range used for finding the offset
of a wavecal observation in the dispersion direction.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@996 fe389314-cf27-0410-b35b-8c050e845b92