Changes to ccos.c:
The calling sequence of xy_extract was modified; this function now extracts
a 2-D array (like ccos.extractband), and it has new arguments pixel_zero,
sdqflags, and the epsilon column.

Changes to accum.py:

Functions pertaining to finding the stims and updating header keywords have
been modified to take pseudo-timetag data (pixel coordinates) as input,
rather than the actual accum image, and these functions call routines in
timetag.py to find the stims.

Changes to calcos.py:

Function extractSpectrum was modified so that for time-tag data there is an
option (currently commented out, however) to pass the corrtag table as the
input to extract1D.

Changes to concurrent.py:

Function findShifts was modified to change the calling sequence for
xy_extract, and to add the extracted band along axis 0 to get the spectrum.
sdqflags is now assigned, but the epsilon column is still left out because
the flat field is not used for FUV and is currently 1 for the wavecal side
of NUV.  (However, epsilon could still be relevant because of the deadtime
correction.)

Function zeroKeywords has been renamed to setShiftKeywords, and it is now
called following the call to applyCorrections.  This still sets the shift2
keywords to zero, but it now sets one or more of the pshift keywords to the
average of the difference between xfull (or yfull) and the nearest integer.
The point of this is that if the OSM drift is much less than a pixel during
an exposure, xfull can have a fractional part that is non-zero and roughly
constant.  Since the xfull coordinates will be rounded to the nearest
integer when binning to the 2-D flt and counts images (and therefore to
the x1d spectrum), there can be a systematic offset of up to a half a pixel.
This offset is now accounted for with the pshift keywords, which are used
by extract.py when computing the associated wavelengths.  This should only
be an issue for NUV, because for FUV the pseudo-random numbers added to
the pixel coordinates will smear this out.

Function deleteCoordinateKeywords was modified to set the PSHIFT[ABC]
keywords to zero for the output lampflash table.

processConcurrentWavecal now returns (avg_dx, avg_dy, pshift_vs_time).

Changes to cosutil.py:

Function delCorrtagWCS was added.  This deletes the WCS keywords in a
corrtag events table header, so the header can be used as a template for
creating a 1-D extracted spectrum (x1d) table.

Offsets avg_dx and avg_dy have been added to the calling sequences of
updateDQArray and flagOutOfBounds.

In returnGTI, the test on the number of elements in the HDU list was
changed from "if len (fd) != 3" to "if len (fd) < 3".  This is to allow
additional extensions in the input file.

Changes to extract.py:

The function extract1D can now accept as input either a corrtag table or the
flt and counts images.  Add function extractCorrtag for extracting from a
corrtag table.

extract1D now calls delCorrtagWCS if the input is a corrtag table.

extractSegment returns the data quality weight array, in addition to the
arrays that it returned previously.  This is a new column in the output
x1d table.

Changes to timetag.py:

Functions initTempcorr, computeThermalParam, findStim, updateStimSum, and
stimKeywords, which find the stims and update header keywords, have been
modified to include the RMS widths of the stims (which are now computed by
findStim).

concurrent.processConcurrentWavecal now returns the average offsets in the
X and Y directions, as well as the pshift (dispersion direction) offsets at
one-second intervals throughout the exposure; these will be 0 and None for
non-tagflash data.  Function appendPshift has been added; this writes the
array of pshift values (at one-second intervals throughout the exposure)
to a new extension with EXTNAME = 'INFO' to the corrtag, flt, and counts
files.  (At the present time, the call to appendPshift is commented out.)

The average offsets (avg_dx, avg_dy) are included in the calling sequence
for doDqicorr, and in that function those values (rounded to integer) are
passed to updateDQArray and flagOutOfBounds.

dopplerRegions was modified to take account of the slope when calculating
the cutoff (boundary) between regions (given by shift_flags) that should or
should not be shifted for the Doppler shift.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@1334 fe389314-cf27-0410-b35b-8c050e845b92