relevant OPRs:  60176, 60197, 60337, 60338, 60339. relevant Trac tickets:  212, 213, 218, 219.
Changes to accum.py:

Rename doHelcorr to initHelcorr, since all it does is compute the radial
velocity and update the V_HELIO keyword.  The heliocentric correction is
actually done by modifying the wavelengths in extract.py.

Change to average.py:

Call cosutil.isProduct and cosutil.modifyAsnMtyp in avgImage to change
the "EXP" part of keyword ASN_MTYP to "PROD", if the output file is a
"product" (which it should be).

Changes to calcos.py:

Rewrite findFirstScience so it returns a tuple of two integers, the
indexes of the first time-tag and first accum science observations.
Modify compareRefFiles and compareSwitches to check time-tag and accum
separately.

In basicCal, pass self.wavecal_info to timetag.timetagBasicCalibration,
so auto and GO wavecals can be treated (mostly) the same as for tagflash
wavecals.

Modify updateShift.  This is now only used for TIME-TAG spectroscopic
data.  The function wavecal.returnWavecalShift now returns both the shift
dictionary and the slope dictionary (the latter is not used here).

Rewrite initialInfo in Association and change Observation, to allow the
filename suffix for ACCUM data to be either "_rawaccum" or "_rawimage";
see PR 60176.

Change allScience and allWavecals to not delete the count rate file.
Delete the removeCountRateFile method.  See PR 60197.

In missingRefFiles(), change min_ver from "1.0" to "2.0" in the dictionary
'temp' (which will be copied to 'ref').

Changes to cosutil.py:

Add functions isProduct and modifyAsnMtyp.  Call these in renameFile to
change ASN_MTYP in any file that is to be a "product" to be "PROD-xxx",
where xxx is unchanged from the value in the raw file, e.g. change
"EXP-FP" to "PROD-FP" and "EXP-RPT" to "PROD-RPT".

Subtract (instead of adding) avg_dx and avg_dy from lx & ly in updateDQArray
and from x0 & y0 in flagOutOfBounds.  This is so that the flagged regions
will be correctly offset to account for OSM shift and/or FPPOS offset.

In writeOutputEvents, include the RAWX and RAWY columns for an FUV corrtag
table.  Also include XFULL and YFULL regardless of whether the wavecal is
tagflash or not.

Changes to extract.py:

Include DQ_OUT_OF_BOUNDS in REALLY_BAD_DQ_FLAGS.  This is so that regions
that are off the detector will be flagged as bad in the DQ_WGT column.

In doExtract(), check for helcorr = "COMPLETE" in addition to "PERFORM",
although this should no longer be necessary.  Move this section (which
modifies the wavelengths) to a point following where the spectrum is
extracted and also following conversion from NET to FLUX.  Both for
converting to flux and for filtering out the atmospheric lines for finding
the spectrum, the wavelengths should be geocentric.

Pass the actual cross-dispersion location to updateExtractionKeywords(),
rather than having that function recompute the location.
Change the order of the NUV segment names to "NUVA", "NUVB", "NUVC".

Changes to fpavg.py:

Call cosutil.isProduct and cosutil.modifyAsnMtyp in __init__ in OutputX1D
to change the "EXP" part of keyword ASN_MTYP to "PROD", if the output file
is a "product" (which it should be).

Include DQ_WGT (a column in x1d tables) in the weight and in the output
x1dsum table.  The value of DQ_WGT in the x1dsum table is the sum of the
input values.

Changes to timetag.py:

In writeImages, include DQ_OUT_OF_BOUNDS in sdqflags.

Add the function updateFromWavecal, which updates columns XFULL & YFULL
based on wavecal info for the case that auto or GO wavecals were used
(as opposed to tagflash wavecals).

Add an initHelcorr function, which computes the radial velocity and
updates the V_HELIO keyword.  Modify doDoppcorr and dopplerCorrection
to remove all references to helcorr.  A significant effect of this
change is that the XDOPP column is no longer corrected for heliocentric
Doppler shift, only orbital Doppler.  This also results in flt and
counts images that are (nominally) the same for time-tag and accum data,
while before they were shifted by the heliocentric correction for
time-tag data but not for accum.

Write the reference locations of the stims to header keywords.
See PR 60231.

Add function setActiveArea, to assign a Boolean array of flags -- one
element for each event -- indicating which events are within the active
area.  This isn't really correct, however, because it's set before doing
the TEMPCORR step.

Changes to wavecal.py:

For non-tagflash wavecals, wavecal.findWavecalShift was adding the
segments together even for FUV data.  The rows of the x1d table (segments
or stripes) should only be added together for NUV data.

In returnWavecalShift, for the case that there was no matching wavecal
(i.e. different FPOFFSET), the nearest wavecal in time could be used,
and a correction is applied depending on the difference in FPOFFSET.
However, the correction was applied to the SHIFT2* keywords as well as
to PSHIFT*, and only the PSHIFT* should have been changed.
Also, the sign of the correction was wrong.

The function returnWavecalShift now computes and returns the slope
as well as the shift; both are dictionaries with the same keys.

In ttFindSpec(), fix the comment about left edge for FUV but bottom edge
for NUV.  slope was gotten but not used, so don't get it.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@6798 fe389314-cf27-0410-b35b-8c050e845b92