Changes to ccos.c:
Functions csum_2d and csum_3d now have binx and biny arguments.
Trac #401, PR 62440

Changes to accum.py:

Function accumBasicCalibration has new argument cl_args.
Trac #394, 401, PR 62433, 62440

Changes to calcos.py:

Command-line arguments --binx, --biny, --compress and --shift were added.
Some of the arguments are now included in variable cl_args, which is passed
to Association instead of separate parameters, and cl_args is also passed to
timetagBasicCalibration and accumBasicCalibration.
Trac #394, 401, PR 62433, 62440

In processWavecal in Calibration, the name of the shift file and obs.info
were added to the calling sequence of wavecal.findWavecalShift.
Trac #394, PR 62433

In Observation, the value of filenames["root"] was added to info, with key
"root".  This is the rootname, but it's based on the name of the file on
disk, which could differ from the value of the rootname keyword if the disk
file has been renamed.  This is to be used by the --shift option as a file
identifier.  Trac #394, PR 62433

Changes to concurrent.py:

Modified avgShift and setShiftKeywords to respectively return and accept
avg_dx and avg_dy that are dictionaries of average shifts, with segment
(stripe) as key, rather than just a value for one segment/stripe.
processConcurrentWavecal now only returns shift1_vs_time, instead of
(avg_dx, avg_dy, shift1_vs_time).  Trac #394, PR 62433

Included the option to use a user-supplied text file specifying shift1
(and shift2, but the latter is currently ignored).  When printing the shifts,
also print the "original" shift1 value, which can differ if shift1 was not
found, or if the user overrode that value with the --shift command-line
option.  Trac #394, PR 62433

Use dispersion.Dispersion instead of cosutil.getTable and cosutil.evalDisp.
Trac #406, PR 62590

Changes to cosutil.py:

Argument shift_file was added to printFilenames.  Trac #394, PR 62433

Delete evalDisp, evalDerivDisp and evalInvDisp (because they're now in
dispersion.py).  Trac #406, PR 62590

New file dispersion.py:

This file contains the Dispersion class for finding the row(s) in the
disptab and evaluating the dispersion relation.  Trac #406, PR 62590

Changes to extract.py:

Use dispersion.Dispersion instead of cosutil.getTable and cosutil.evalDisp.
Trac #406, PR 62590

Changes to findshift1.py:

Minimize chi square, rather than using cross correlation, to find the shift
in the dispersion direction.  Save the "original" shift1 value (which might
have not been well determined), in addition to the shift1 that will be used
(and might be a default value or based on an average of the shifts of the
other NUV stripes).  Trac #394, PR 62433

Changes to timetag.py:

Function timetagBasicCalibration has new argument cl_args.
Trac #394, 401, PR 62433, 62440

updateFromWavecal now returns just shift1_vs_time, instead of (avg_dx, avg_dy,
shift1_vs_time), and avg_dx is no longer computed.  Trac #394, PR 62433

Some of the elements in cl_args are passed to writeCsum, and that function
permits binning and/or compression of the "calcos sum" image.
Trac #401, PR 62440

Use dispersion.Dispersion instead of cosutil.getTable, cosutil.evalDisp,
and cosutil.evalDerivDisp.  Trac #406, PR 62590

Changes to wavecal.py:

Added info and shift_file to the calling sequence of findWavecalShift.  In
that function, override shift1 if the user supplied a shift_file.
When printing the shifts, also print the "original" shift1 value.
Trac #394, PR 62433


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@8026 fe389314-cf27-0410-b35b-8c050e845b92