Changes to ccos.c:
Added function dq_or.  Trac #302, PR 61305
Added function bin2d.  Trac #328, PR 61783, AI 090409-8

Changes to accum.py:

Implemented calibration for acq/image data.  Changed the name of writeImset
to appendImset, and added function writePrimaryHDU.
Trac #328, PR 61783, AI 090409-7, 090409-8

Changes to calcos.py:

The following keywords in mergeKeywords were added to the list that will be
merged from segment A to B and from B to A:  npha_[ab] (name change from
pha_bad[ab]), tbrst_[ab], nbrst_[ab], tbadt_[ab], nbadt_[ab], nout_[ab].
Trac #408, PR 62666, AI 090409-9

Changes to concurrent.py:

Modified findShifts for the case that no spectrum was found for a given flash.
spec_found is now set to False for all segments (stripes), rather than printing
"no spectrum found" and skipping the remainder of the code in the loop over
flashes.  The rest of the loop needed to be executed in order to write the
info (i.e. spec not found) to the lampflash table, and to avoid getting the
list of shift1 values out of synch with chi_square and n_deg_freedom.
Trac #394, PR 62433

Changes to cosutil.py:

Function determineLivetime was moved from timetag.py to cosutil.py, because
it's now called in both timetag.py and accum.py.
An imset argument was added to function getInputDQ, with a default value
of 1 so the call in timetag.doDqicorr does not need to be changed.  This
argument is needed for acq/image data.
Trac #328, PR 61783, AI 090409-7

Changes to extract.py:

In extractSegment, the entire background array for NUV data was being boxcar
smoothed, but there are now buffer zones with zero count rate on the left and
right ends, and those buffer zones should not be included in the smoothing
because they will have systematically lower values (i.e. zero) than the real
background.  The smoothing is now done only over 1024 pixels, offset by
100 pixels (x_offset) and -shift1[abc].  Trac #411, PR 61914

The following keywords were added to the list in concatenateFUVSegments
that will be copied from segment B to segment A:  npha_b (name change from
pha_badb), tbrst_b, nbrst_b, tbadt_b, nbadt_b, nout_b.  Also, keyword
nbadevnt is set to the sum of the values of that keyword in the segment A
and B headers.
Trac #408, PR 62666, AI 090409-9

In extractSegment, call ccos.dq_or instead of looping over rows of dq_ij
to get DQ_i.  Trac #302, PR 61305

Changes to timetag.py:

timetagBasicCalibration and updateFromWavecal were modified to include
the name of the shift file (if not None) when printing the name(s) of
the wavecal files to the trailer.  updateFromWavecal also includes the
shift file name in the WAVECALS primary header keyword.  Trac #395, PR 62434

In timetagBasicCalibration, a list of headers is now created.  If the
input file is an ACCUM exposure, this list will contain all the headers
from the input file:  primary, sci, err, dq.  If the input file is a
TIME-TAG exposure, this list will contain four headers (for compatibility),
but they will be the primary, events, and two more copies of the events
header.  This list will be passed to writeNull, writeImages and makeImage,
which can then write different headers for the sci, err and dq extensions
if the input is ACCUM.  Trac #407, PR 62665, AI 090409-11

Function countBadEvents was added.  This determines the number of events
and the amount of time lost due to FUV bursts and to the badttab.  It also
determines the number of events that are outside the FUV active area.  These
values are saved in keywords.
Function updateGlobrate was modified to change the keyword name from
globrate to globrt_a or globrt_b for FUV data (or leave it as globrate
for NUV data).
Function filterByPulseHeight was modified to change the keywords for the
number of events rejected for pulse height out of bounds from pha_bad[ab]
to npha_[ab].
Trac #408, PR 62666, AI 090409-9

Added code to the function saveNewGTI to create a new GTI bintable header/data
unit and append it to the output HDU list.  Incremented extver for that
extension to be one larger than any other GTI table already in the raw file
(but we expect only one).  Trac #409, PR 62667, AI 090409-12

Function determineLivetime was moved from timetag.py to cosutil.py.
Trac #328, PR 61783, AI 090409-7


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@8065 fe389314-cf27-0410-b35b-8c050e845b92