All these changes were made to support imaging wavecals, although the "support" in some cases just means calibrating as an image. Trac #367, PR 62091
In __init__ for Association, count exp_type = EXP_CALIBRATION as "science"
for concat_info.

In findFirstScience, include exp_type = EXP_CALIBRATION in the checks
for the first wavecal.

In missingRefFiles, only look for lamptab and wcptab for spectroscopic
wavecals.

In globalSwitches, also set global_switches["science"] to PERFORM if
exp_type = EXP_CALIBRATION.

In copySptFile, also check for exp_type = EXP_CALIBRATION, and use that
in preference to 'fallback'.

In __init__ for Observation, rewrite the section for checking obstype,
exptype, tagflash, and memtype.

In getHeaderInfo, close the input file after getting the headers.

In checkImSpecSwitches, reset photcorr to OMIT if the aperture is either
"WCA" or "FCA".

In allScience, process exp_type = EXP_CALIBRATION as science data.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@8165 fe389314-cf27-0410-b35b-8c050e845b92