Changes to calcos.py:
In __init__ for Observation, allow the input to be a corrtag file as long
as the input and output directories are not the same.  That is, an output
directory doesn't have to be specified if the input files are in another
directory.  Function expandDirectory was added to calcos.py to support this,
so the directory names will be complete paths and can be directly compared.
Trac #417, PR 62895

Changes to concurrent.py:

Functions shift1Corr, shift2corr and setShiftKeywords have been modified
to use column YCORR (variable eta) instead of YFULL (eta_corr) for assigning
values to shift_flags, the array of flags that indicate which events are
within the active region.  Most of those changes were for consistency of
notation; the one that counts is shift1Corr, because at the time that that
function is called there is (usually) a difference between YCORR and YFULL.
Trac #437, PR 63346

In avgShift, if there are no flashes or if the exposure time is zero,
assign initial (0) values in avg_dx and avg_dy.  This was done because
setShiftKeywords assumes that these dictionaries have the same keys as
other dictionaries of parameter values.  Trac #438, PR 63347

In processConcurrentWavecal, set LAMPUSED to NONE (indicating that the lamp
was off) before writing an empty lampflash table if no lamp flash is found.
Trac #425, PR 62997

Changes to cosutil.py:

In minmaxDoppler, compute the minimum and maximum Doppler shifts if doppcorr
is either PERFORM or COMPLETE, not just PERFORM.  Trac #417, PR 62895


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@8237 fe389314-cf27-0410-b35b-8c050e845b92