Changes to findshift1.py:
Function setShift1 now checks whether the key (in the calling sequence) is
present in shift1_dict, and if not, the function returns without setting
anything.  Function findShift now checks that the coeff array (returned by
cosutil.fitQuadratic) is not None.  Trac #394, PR 62433

Change to wavecal.py:

In findWavecalShift, a test was added to check that there actually is an
"FUVA" in the SEGMENT column before setting override_segment_B to True.
Trac #394, PR 62433


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@8543 fe389314-cf27-0410-b35b-8c050e845b92