Change to accum.py:
Function doDqicorr was modified to pass a dictionary instead of a list to
cosutil.updateDQArray, and a dummy doppler_boundary was added to the
calling sequence.  Trac #524, PR 64722

Changes to concurrent.py:

Function shift1corr now uses '<' insead of '<=' for region[1].
Trac #524, PR 64722

Method setRegions in NUVConcurrentWavecal was modified to include the slope
when finding the Y location, i.e. the Y value is now at the middle of the
detector rather than at the left edge.  Previously it wasn't consistent with
the code for selecting regions in timetag.py.  Trac #524, PR 64722

Function findShifts now prints the shift that was actually measured and
fp_pixel_shift from the lamptab, rather than just the total shift (the
measured shift plus fp_pixel_shift).  Trac #525, PR 64723

Changes to cosutil.py:

Functions updateDQArray, flagOutOfBounds and flagOutsideActiveArea were
modified to accept a dictionary for the minimum and maximum wavecal shifts
instead of just a list (most changes were to updateDQArray).
A new argument doppler_boundary was added to updateDQArray.  FUV data are
now handled in two slices, one for the PSA region and one for the WCA, and
the doppler offset (minmax_doppler) is only applied in the PSA region.
Trac #524, PR 64722

Changes to timetag.py:

Functions nuvPsaRegions and nuvWcaRegions were broken into two function calls,
nuvPsaBoundaries (or nuvWcaBoundaries) and flagsFromBoundaries.
nuvPsaBoundaries and nuvWcaBoundaries are now called in getWavecalOffsets,
and the variable returned by getWavecalOffsets is a dictionary of minimum
and maximum shifts, called minmax_shift_dict, with stripe as the key.
Trac #524, PR 64722

A new function psaWcaBoundary was added, extracted from fuvDopplerRegions
and modified to work for either FUV or NUV.  fuvDopplerRegions calls
psaWcaBoundary.  psaWcaBoundary is also called in doDqicorr, and the location
of the boundary (doppler_boundary) is passed to cosutil.updateDQArray.
Trac #524, PR 64722

Change to wavecal.py:

Function findWavecalShift now prints the measured shift and fp_pixel_shift
from the lamptab (or 0).  Trac #525, PR 64723


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@8916 fe389314-cf27-0410-b35b-8c050e845b92