Change to extract.py:
xd_nominal is now assigned a value regardless of whether the exposure
is a wavecal or not, and it is assigned in both extractSegment and in
extractCorrtag.
Trac #575, PR 65353

Change to xd_search.py:

In function findPixelNumber, the indices i0 or i1 could extend beyond
the range of the data, and in that case they are now truncated.
Trac #575, PR 65353


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@9800 fe389314-cf27-0410-b35b-8c050e845b92