Changes to calcos.py:
A new command-line option --raw was added.  If specified, and if a csum file
is to be created, the raw pixel coordinates (rather than thermally and
geometrically corrected) will be used to make the csum image array.
Trac #581, PR 65548

Changes to timetag.py:

The calling sequence of writeCsum was modified.  The events data array and
the raw_csum_coords flag are now passed as arguments.  Note that this is
relevant only for FUV, so the writeCsum function in accum.py was not modified,
as it is only called for acq/image data.
Trac #581, PR 65548

In function recomputeExptime, the code to update keyword EXPTIMEA or EXPTIMEB
was only executed if the recomputed exposure time differed from the original
exposure time.  Furthermore, recomputeExptime was only called for TIME-TAG
data, but the segment-specific exposure time keywords need to be updated for
any FUV exposure.  The code to update these keywords has been moved out of
recomputeExptime, and a separate test has been added to check that the exposure
is FUV.  Trac #556, PR 65023

In function initTempcorr, counts1 and counts2 were supposed to be the sums
of the stim counts over all time intervals, but they were only the counts
within the last time interval.  sumstim[0] and sumstim[5] have the accumulated
counts, and these values are now assigned to total_counts1 and total_counts2
respectively.  Trac #579, PR 65502


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@9878 fe389314-cf27-0410-b35b-8c050e845b92