Changes to calcos.py:
The __init__ method in Association was modified (for FUV) to include a list
of the corrtag file names in merge_kwds, so segment-specific keywords will be
copied from the corrtag FUVA headers to FUVB headers and vice-versa.
In Calibration.extractSpectrum, function updateCorrtagKeywords in extract.py
is called to copy extraction-location keywords from the flt header to the
corrtag header.  Trac #586, PR 65692

Changes to calcosparam.py:

Variable segment_specific_keywords is a new list of segment-specific keywords.
This is used in calcos.mergeKeywords and extract.concatenateFUVSegments,
to avoid having independent lists in those modules.  Trac #586, PR 65692

Changes to extract.py:

The list segment_specific_keywords is used in concatenateFUVSegments instead
of having a local copy of the keywords to be copied from segment A to B.
Function updateCorrtagKeywords was added; this copies extraction-location
keywords from an flt header to a corrtag header.  Trac #586, PR 65692

Changes to timetag.py:

Function mkHeaders was added.  For TIME-TAG input, this creates the list of
four headers that will be used when creating the extensions for the flt and
counts files.  Trac #587, PR 65693


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@10045 fe389314-cf27-0410-b35b-8c050e845b92