Changes to concurrent.py:
In setShiftKeywords, a test was added to see whether there were any elements
in xi before taking its mean value.  Trac #591, PR 65793

Changes to extract.py:

In extractSegment, a test was added to see whether found_locn is None, and a
comment pertaining to this was added to extractCorrtag.
Trac #591, PR 65793

Changes to timetag.py:

In updateFromWavecal, a test was added on xi_diff to make sure it has any
elements before taking its mean value.  In getWavecalOffsets, a test was
added on xdiff_subset to make sure it has any elements before finding its
minimum and maximum values.
Trac #591, PR 65793

Function mkHeaders was assigning EXTNAME=SCI and EXTVER=1 in the first
extension header.  This was OK when the header was used as a template for
the output flt and counts files, but this header was also the output header
for the corrtag table, and in the latter EXTNAME should be EVENTS, not SCI.
The code to assign these two values in the first extension header has been
deleted; EXTNAME is assigned in makeImageHDU when the output header/data unit
is created.  Trac #587, PR 65693

Changes to xd_search.py:

In xdSearch, y_locn is checked to see that it is not None before adding
the zero point to it.  Trac #591, PR 65793


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@10081 fe389314-cf27-0410-b35b-8c050e845b92