Changes to cosutil.py:
Function cmpVersion now appends zero or more "0" elements to each list
(minv, phdrv, calv) after splitting the version strings on ".", until all
those lists have the same length.  Function cmpPart was greatly simplified.
Function splitIntLetter was added, to split for example "13b" to "13" and "b".
Variable phdr_ver was changed to vcalcos (the name of the keyword) in
functions findRefFile and cmpVersion, just to make the code more readable.
In cmpVersion, values 0 and 1 for boolean variables were changed to False
and True respectively.
Trac #620, PR 66250

Changes to extract.py:

Function updateArchiveSearch was modified to update keywords PLATESC and
SPECRES in the primary header, using values read from the SPWCSTAB.
Trac #619, PR 66249

Changes to timetag.py:

Functions filterByPulseHeight and checkPulseHeight were modified to include
OPT_ELEM as a selection criterion for the PHATAB (pulse-height parameters
table), if that column is present in the PHATAB.  The calling sequence for
filterByPulseHeight was modified to include the info dictionary.
Trac #618, PR 66221


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@10484 fe389314-cf27-0410-b35b-8c050e845b92