Changes to ccos.c:
New function pha_check was added, for doing the pulse-height correction
(i.e. flagging the DQ array) using a pair of images rather than a table.
Trac #628, PR 66626

Doc strings were modified in cosutil.py, dispersion.py, extract.py,
findshift1.py, fpavg.py, shiftfile.py, splittag.py, spwcs.py, timetag.py,
wavecal.py, x1d.py, xd_search.py.
Trac #616, PR 66158

Changes to calcos.py:

Function missingRefFiles was changed to include "phafile" in the list of
reference files, and for TIME-TAG data to use that instead of "phatab" if
"phafile" is not "N/A".
Trac #628, PR 66626

Changes to concurrent.py:

Method updateOutFlash was renamed to copySegAtoSegB, and new methods
setSegBtoZero and miscSegB were added.  Attribute segment_A_present was added,
a flag to indicate (for G140L data) whether there was a raw file for segment A.
Method copySegAtoSegB was modified to correct a bug in the section for setting
the wavecal shift based on a user-supplied value.  The element in the shift1
list was the shift itself, but it should have been a dictionary with segment
as the key.
Trac #465, PR 63864

Changes to findshift1.py:

A new fp argument was added to setShift1.  If fp=True (the default), then
fp_pixel_shift will be subtracted from the input shift1 value before saving
in shift1_dict.  Regardless of the value of fp, fp_pixel_shift will be
subtracted from shift1 in the call to computeNormalization.
Trac #465, PR 63864

Changes to getinfo.py:

Function getRefFileNames was modified to get keyword phafile as well as
phatab.  If either keyword is missing, its value in the reffiles dictionary
will be "N/A".
Trac #628, PR 66626

Changes to timetag.py:

A new filterPHA function was added.  For a TIME-TAG exposure, this function
will be called if the input header specified a PHAFILE; otherwise, the
original filterByPulseHeight function will be called, using a PHATAB.
Trac #628, PR 66626

Changes to wavecal.py:

In function findWavecalShift, for the case of G140L, FUVB data with no FUVA
data, shift1 is now set to zero.  In the call to setShift1 to set the value
to 0, the new fp argument is set to False so fp_pixel_shift will not be
subtracted.
Trac #465, PR 63864


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@11259 fe389314-cf27-0410-b35b-8c050e845b92