Changes to accum.py:
For the case that the input SCI image either has no data block or no counts
in the image, accumBasicCalibration was calling writeNull to write the flt
and counts files (but not corrtag) and returning.  This was modified to
create the corrtag file (but not the flt and counts, at least right away)
and then call timetag.timetagBasicCalibration as usual.  The latter creates
the flt and counts files, and it may also create the csum file.  The writeNull
function is no longer called, and it has been deleted.
Trac #635, PR 66908

Changes to calcos.py:

The main module was changed to delete the command-line argument --only_raw_csum
and to add --only_csum.  Function prtOptions now shows option --only_csum.
The argument list for calcos was changed to replace `only_raw_csum` with
`only_csum`, and the comments were updated.  The entry in the cl_args
dictionary was changed from "only_raw_csum" to "only_csum"; this is used in
updateCombineFlt, updateCombineX1d and resetSwitches, and in timetag.py.
Trac #657, PR 67439

Changes to concurrent.py:

Function processConcurrentWavecal now returns both an array of times and
an array of corresponding shift1 values (instead of just the latter).
Function shift1VsTime was rewritten.  The previous version attempted to
integrate over one-second intervals, but it didn't work correctly.  This
version samples the shift in the dispersion direction.
Function getInterval was modified to explicitly set i1 to the length of the
time column if the current flash is the last one.  ccos.range should have
returned this value, but sometimes it has problems with roundoff.
Trac #645, PR 67204

Changes to cosutil.py:

Function timelineTimes was added.
Trac #645, PR 67204

Function updateDQArray was modified.  When looping over the keys in
minmax_dict (copied from minmax_shift_dict), the keys are now sorted before
the loop.  upper_y_s was truncated at the upper edge of the image as follows:
    if upper_y_s >= dq_shape[0]:
        upper_y_s = dq_shape[0] - 2
This was changed to:
    if upper_y_s > dq_shape[0]:
        upper_y_s = dq_shape[0]
Trac #640, PR 67115

Changes to spwcs.py:

Method addKeywords was modified to put the WCSAXES keywords just after GCOUNT,
instead of just before CTYPE1.  It's a FITS requirement that WCSAXES precede
all other WCS keywords, but CTYPE1 isn't always present, and it doesn't have
to be the first WCS keyword.
Method deleteKeywords for image data was modified to also delete keywords
that are specific to a different type of header from the current one.  For
example, TCTYP7 will be deleted from an flt or counts header.
Trac #639, PR 67114

New file orbit.py:

This file contains the code to read the orbital parameters from the spt file
and to compute the position and velocity of HST.
Trac #645, PR 67204

New file timeline.py:

This file contains the code to create a TIMELINE extension for a corrtag
file, and to populate the columns of this table.
Trac #645, PR 67204

Changes to timetag.py:

Function concurrent.processConcurrentWavecal or updateFromWavecal will
now only be called if wavecorr is perform.  This is so that when processing
a corrtag file, wavecal processing will not be repeated unless the user has
explicitly set wavecorr to perform.
Trac #661, PR 67473

Entry "only_raw_csum" in cl_args was changed to "only_csum".  This is used
in function timetagBasicCalibration.
Trac #657, PR 67439

Function timeline.createTimeline is called to create and append the timeline
bintable extension.  Function appendShift1 (which was defined but not actually
called) has been deleted.
Function updateFromWavecal calls cosutil.timelineTimes to get the array
of times at which to evaluate shift1.  updateFromWavecal now returns both
the array of times and the array of shift1 values.
The file handle ofd for the output corrtag file is now passed to writeNull,
and that function now appends a dummy timeline table.
Trac #645, PR 67204

Function writeCsum now uses a test on obsmode rather than whether the pha
column exists in order to determine the dimension of the csum file for
FUV TIME-TAG data.  For the case that the raw file contained no data,
writeCsum had been writing files that were only 2-D (for FUV TIME-TAG data).
Function computeWavelengths now checks whether the input `events` is None,
and if so, returns without doing anything.
Trac #635, PR 66908


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@11949 fe389314-cf27-0410-b35b-8c050e845b92