Changes to timetag.py:
Function xyWalk was modified so it subtracts the correction from XCORR
and YCORR; previously the correction was added.
Trac #702, PR 68209

Changes to splittag.py:

In function splitOneTag, nevents (number of events) is passed as a new
argument to updateKeywords.  For FUV data, function getInfo gets the exposure
time from either keyword EXPTIMEA or EXPTIMEB, depending on segment.  Function
updateKeywords has a new nevents argument, and it now updates this value and
the exposure time with (for FUV) segment-specific keywords.
Trac #707, PR 68355

Changes to x1d.py:

Function makeFltCounts was modified to get the exposure time from either
keyword EXPTIMEA or EXPTIMEB, for FUV data, depending on segment.
Trac #707, PR 68355


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@12870 fe389314-cf27-0410-b35b-8c050e845b92