The following files were changed to update some Python syntax that is now considered deprecated: average.py, calcos.py, concurrent.py, cosutil.py, dispersion.py, extract.py, findshift1.py, fpavg.py, getinfo.py, phot.py, shiftfile.py, splittag.py, timetag.py, wavecal.py, xd_search.py, x1d.py Specifically, the `raise` statement uses parentheses (instead of a comma) after the exception class, use the `key` argument to the built-in sort function rather than the `cmp` argument, use the `sorted` function to get a sorted list of dictionary keys rather than assuming that the keys() method of a dictionary returns a list.  Other similar changes (listed individually below) are to replace the use of has_key() for a dictionary and to use isinstance() instead of type(). Trac #718, PR 68538
Changes to ccos.c:

The smoothbkg function was modified by adding an optional argument, an array
of flags (0 is good, 1 is bad).  The edge extrapolation was changed from
using the value of the nearest pixel to shrinking the smoothing box size.
Trac #736, PR 68713

Changes to calcos.py:

The following functions were modified to remove use of has_key for a
dictionary:  checkNumerix, updateCombineFlt, updateCombineX1d,
globalSwitches, fixRelMvReq, overrideSwitch, combineToProduct,
combineFlt, combineAllX1D.
Trac #718, PR 68538

Method compareConfig was modified to allow exposures with different values
of cenwave to be included in an association.  In processWavecal, cenwave
was added to the calling sequence of wavecal.storeWavecalInfo.
Trac #720, PR 68550

A new command-line flag --nofind was added.  The main module and prtOptions
were modified.
Trac #719, PR 68545

Changes to concurrent.py:

The r_cmp function in setRegions in NUVConcurrentWavecal was replaced by
r_key, and this is specified by keyword `key` instead of `cmp`.  Function
r_cmp returns the second element (b_spec, ignoring segment) of its argument.
Trac #718, PR 68538

Changes to cosutil.py:

Function combineStat was modified to remove the use of has_key for a
dictionary.
Trac #718, PR 68538

Functions findColumn and findRefFile were change to replace calls to "type"
with calls to "isinstance".
Trac #718, PR 68538

Function printMsg was changed to use 'print (message)' instead of
'print message'.
Trac #718, PR 68538

Changes to dispersion.py:

Dispersion.__init__ was modified to not use has_key for a dictionary.
Trac #718, PR 68538

Changes to extract.py:

The section for smoothing the background in extractSegment and extractCorrtag
was modified to use ccos.smoothbkg instead of convolve.boxcar.  The call to
boxcar used mode='nearest' for endpoint extrapolation (as did the previous
version of smoothbkg), but the revised smoothbkg effectively shrinks the box
size near the endpoints, which is more appropriate for background smoothing.
For columns flagged as bad over both background regions, those are no longer
set to zero in the smoothed array; instead, the original background values
are copied back to those parts of the background array after smoothing.
When truncating j at the upper end of the array size, axis_length is used
rather than axis_length-1, since j is the upper limit of a slice.  A new
function, excludeAllBad, is now called in the background smoothing section,
so that the smoothing does not include regions for which the DQ is bad for
an entire column.
Trac #736, PR 68713

After calling xd_search.xdSearch in extractSegment and extractCorrtag, the
error estimate and FWHM of the cross-dispersion profile are now printed out
as information to the user.  These are printed just after the found and
nominal locations.
Trac #717, PR 68529

Changes to wavecal.py:

The import statement for convolve was modified, because the name has been
changed to stsci.convolve.  For backward compatibility, try/except is used,
checking for ImportError for stsci.convolve.
Trac #735, PR 68712

Function interpolateWavecal was modified to use "if key in shift_dict1"
rather than "if shift_dict1.has_key (key)".
Trac #718, PR 68538

The central wavelength was added to function storeWavecalInfo,
returnWavecalShift, selectWavecalInfo.  If there is no matching row in
returnWavecalShift, the central wavelengths also have to match before we
will try to correct for a difference in fpoffset.
Trac #720, PR 68550

Function cmpTime was replaced by keyTime (which returns the element of
the input dictionary with key "time"), and this function is passed with
keyword "key" to the sort function in storeWavecalInfo.
Trac #718, PR 68538

Changes to xd_search.py:

The import statement for convolve was modified, because the name has been
changed to stsci.convolve.  For backward compatibility, try/except is used,
checking for ImportError for stsci.convolve.
Trac #735, PR 68712

The findPeak function now calls new function findFwhm to get the FWHM of
the cross-dispersion profile, and this value is included in the tuple
returned by xdSearch.
Trac #717, PR 68529


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/calcos@13311 fe389314-cf27-0410-b35b-8c050e845b92